<==== Mercurial data ====>
Branch: vms_4.0

+++ vms/libs/nx_sdk/src/nx/sdk/analytics/helpers/object_metadata_packet.cpp
@@ -34,12 +34,12 @@ int64_t ObjectMetadataPacket::durationUs() const
 
 int ObjectMetadataPacket::count() const
 {
-    return m_objects.size();
+    return (int) m_objects.size();
 }
 
 const IObjectMetadata* ObjectMetadataPacket::at(int index) const
 {
-    if (index < 0 || index >= m_objects.size())
+    if (index < 0 || index >= (int) m_objects.size())
         return nullptr;
 
     return m_objects[index];
